+++ NEWS.md
@@ -1,4 +1,11 @@
 
+# devoutsvg 0.2.0  2020-01-20
+
+* Removed old `{devoutsvg-0.1.0}` api for pattern integration.
+* Instead, use `pattern_list` object to define patterns and filters to use
+  for specific colours
+* Update default width and height to to match `svglite`
+
 # devoutsvg 0.1.1  2020-01-19
 
 * Added support for patterns with filters